+++ deepface/detectors/FaceDetector.py
@@ -9,7 +9,7 @@ from deepface.detectors import (
     MtcnnWrapper,
     RetinaFaceWrapper,
     MediapipeWrapper,
-    Yolov8nfaceWrapper,
+    Yolov8faceWrapper,
 )
 
 
@@ -23,7 +23,9 @@ def build_model(detector_backend):
         "mtcnn": MtcnnWrapper.build_model,
         "retinaface": RetinaFaceWrapper.build_model,
         "mediapipe": MediapipeWrapper.build_model,
-        "yolov8n-face": Yolov8nfaceWrapper.build_model,
+        "yolov8-lite-t": Yolov8faceWrapper.build_model("yolov8-lite-t"),
+        "yolov8-lite-s": Yolov8faceWrapper.build_model("yolov8-lite-s"),
+        "yolov8n": Yolov8faceWrapper.build_model("yolov8n"),
     }
 
     if not "face_detector_obj" in globals():
@@ -63,7 +65,9 @@ def detect_faces(face_detector, detector_backend, img, align=True):
         "mtcnn": MtcnnWrapper.detect_face,
         "retinaface": RetinaFaceWrapper.detect_face,
         "mediapipe": MediapipeWrapper.detect_face,
-        "yolov8n-face": Yolov8nfaceWrapper.detect_face,
+        "yolov8-lite-t": Yolov8faceWrapper.detect_face,
+        "yolov8-lite-s": Yolov8faceWrapper.detect_face,
+        "yolov8n": Yolov8faceWrapper.detect_face,
     }
 
     detect_face_fn = backends.get(detector_backend)
